requirements.txt
Django==3.1
APScheduler==3.6.3
celery==4.4.7
django-apscheduler==0.5.0
django-celery-beat==2.1.0
flower==0.9.5
pymongo==3.11.0
PyMySQL==0.10.1
PyYAML==5.3.1
redis==3.5.3
mysqlclient==2.0.1
influxdb==5.3.0

